requirements.txt
llama-index>=0.10.0
requests
pandas
transformers
sentence-transformers
python-dotenv
sqlalchemy
flask
flask-cors
torch
